refactor: centralize session utils and remove unused websocket
