+++ src/api/clients.js
@@ -1,21 +1,10 @@
 import axios from "axios";
+import { getAccessToken, clearTokens } from "@/utils/sessionUtils";
 
 // ─────────────────────────────────────────────────────────────
 // 🌐 Base URL de API (Django)
 // ─────────────────────────────────────────────────────────────
 const API_BASE_URL = import.meta.env.VITE_API_BASE_URL || "http://localhost:8000/api/v1";
-
-// ─────────────────────────────────────────────────────────────
-// 🧠 Helpers para tokens
-// ─────────────────────────────────────────────────────────────
-const getAccessToken = () => sessionStorage.getItem("accessToken");
-const getRefreshToken = () => sessionStorage.getItem("refreshToken");
-
-const clearTokens = () => {
-  sessionStorage.removeItem("accessToken");
-  sessionStorage.removeItem("refreshToken");
-  window.dispatchEvent(new Event("sessionExpired"));
-};
 
 // ─────────────────────────────────────────────────────────────
 // 🔧 Factory de clientes con configuración fija (Django)
@@ -48,9 +37,4 @@ export const djangoApi = createApiClient();
 // ─────────────────────────────────────────────────────────────
 // 📤 Exportación utilitaria
 // ─────────────────────────────────────────────────────────────
-export {
-  djangoApi as axiosInstance,
-  getAccessToken,
-  getRefreshToken,
-  clearTokens,
-};
+export { djangoApi as axiosInstance };
