+++ src/hooks/useSession.js
@@ -1,5 +1,5 @@
 import { useEffect } from "react";
-import { getAccessToken, clearTokens } from "../api/clients";
+import { getAccessToken, clearTokens } from "../utils/sessionUtils";
 import { isJwtExpired } from "../utils/jwtUtils";
 
 /**
